% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputef.R
\name{aldknni}
\alias{aldknni}
\title{aldknni}
\usage{
aldknni(fname, 
        min_coverage=0,
        min_allele_frequency=0.0001,
        max_missingness_rate_per_locus=1.00,
        pool_sizes=c(100),
        min_depth_below_which_are_missing=1,
        max_depth_above_which_are_missing=1000000,
        frac_top_missing_pools=0.0,
        frac_top_missing_loci=0.0,
        min_loci_corr=0.9,
        max_pool_dist=0.1,
        min_l_loci=1,
        min_k_neighbours=1,
        restrict_linked_loci_per_chromosome=TRUE,
        n_reps=5,
        n_threads=2,
        fname_out_prefix="")
}
\arguments{
\item{fname}{name of the genotype file to be imputed in uncompressed vcf, sync or allele frequency table format. See genotype format details below.}

\item{min_coverage}{minimum coverage per locus, i.e. if at a locus, a pool falls below this value (does not skip missing data, i.e. missing locus has a depth of zero), then the whole locus is omitted. Set this to zero if the vcf has been filtered and contains missing values, i.e. \code{./.} or \code{.|.}. \link{Default=0}}

\item{min_allele_frequency}{minimum allele frequency per locus, i.e. if at a locus, a pool has all its alleles below this value and/or above the additive complement of this value (skipping missing data), then the entire locus is omitted. \link{Default=0.0001}}

\item{max_missingness_rate_per_locus}{maximum fraction of pools missing per locus, i.e. if at a locus, there were more pools missing than the coverage dictated by this threshold, then the locus is omitted. \link{Default=1.00}}

\item{pool_sizes}{vector of pool sizes, i.e. the number of individuals included in each pool, or can be set to an arbitrarily large value like 100 for individual polyploids or if allele frequency estimates are expected to be accurate. \link{Default=100}}

\item{min_depth_below_which_are_missing}{minimum depth at which loci with depth below this threshold are set to missing. Set to one if the input vcf has already been filtered and the loci beyond the depth thresholds have been set to missing, otherwise set to an integer above zero. \link{Default=1}}

\item{max_depth_above_which_are_missing}{maximum depth at which loci with depth above this threshold are set to missing. Set to some large arbitrarily large value (e.g. 1000000) if the input vcf has already been filtered and the loci beyond the depth thresholds have been set to missing, otherwise set to an integer above zero. \link{Default=1000000}}

\item{frac_top_missing_pools}{fraction of pools with the highest number of missing loci to be omitted. Set to zero if the input vcf has already been filtered and the loci beyond the depth thresholds have been set to missing, otherwise set to a decimal number between zero and one. \link{Default=0.0}}

\item{frac_top_missing_loci}{fraction of loci with the highest number of pools with missing data to be omitted. Set to zero if the input vcf has already been filtered and the loci beyond the depth thresholds have been set to missing, otherwise set to an decimal number between zero and one. \link{Default=0.0}}

\item{min_loci_corr}{Minimum correlation (Pearson's correlation) between the locus requiring imputation and other loci deemed to be in linkage with it. Ranges from 0.0 to 1.0. If using the default value with is NA, then this threshold will be optimised to find the best value minimising imputation error. \link{Default=NA}}

\item{max_pool_dist}{Maximum genetic distance (mean absolute difference in allele frequencies) between the pool or sample requiring imputation and pools or samples deemed to be the closest neighbours. Ranges from 0.0 to 1.0. If using the default value with is NA, then this threshold will be optimised to find the best value minimising imputation error. \link{Default=NA}}

\item{min_l_loci}{Minimum number of linked loci to be used in estimating genetic distances between the pool or sample requiring imputation and other pools or samples. Minimum value of 1. \link{Default=1}}

\item{min_k_neighbours}{Minimum number of k-nearest neighbours of the pool or sample requiring imputation. Minimum value of 1. \link{Default=1}}

\item{restrict_linked_loci_per_chromosome}{Restrict the choice of linked loci to within the chromosome the locus requiring imputation belong to? \link{Default=TRUE}}

\item{n_reps}{Number of replications for the optimisation for the minimum loci correlation, and/or maximum genetic distance. Minimum value of 1. \link{Default=5}}

\item{n_threads}{number of computing threads or processor cores to use in the computations. \link{Default=2}}

\item{fname_out_prefix}{prefix of the output files including the \href{#allele-frequency-table-csv}{imputed allele frequency table} (\verb{<fname_out_prefix>-<time>-<random_id>-IMPUTED.csv}). \link{Default="" which will use the name of the input file as the prefix including the path}}
}
\value{
imputed allele frequency table with the following filename: \verb{<fname_out_prefix>-<time>-<random_id>-IMPUTED.csv}. Additionally, a sync file is generated if the input was an uncompressed vcf.
}
\description{
Adaptive linkage-informed k-nearest neighbour imputation of allele frequencies
}
\details{
This is an attempt to extend the \href{https://doi.org/10.1534/g3.115.021667}{LD-kNNi method of Money et al, 2015, i.e. LinkImpute}, which was an extension of the \href{https://doi.org/10.1093/bioinformatics/17.6.520}{kNN imputation of Troyanskaya et al, 2001}.
Similar to LD-kNNi, linkage disequilibrium (LD) is estimated using Pearson's product moment correlation per pair of loci, which is computed per chromosome by default, but can be computed across the entire genome.
We use the mean absolute difference/error (MAE) between allele frequencies among linked loci as an estimate of genetic distance between samples.
Fixed values for the minimum correlation to identify loci used in distance estimation, and maximum genetic distance to select the k-nearest neighbours can be defined.
Additionally, minimum number of loci to include in distance estimation, and minimum number of nearest neighbours can be set.
Moreover, all four parameters can be optimised, i.e. the minimum correlation and/or maximum distance and/or minimum number of loci and/or minimum number of nearest neighbours which minimises the MAE between predicted and expected allele frequencies after simulating 10\% missing data are identified.
The allele depth information (AD), i.e. the unfiltered allele depth which includes the reads which did not pass the variant caller filters are used to calculate allele frequencies.
If the GT field is present but the AD field is absent, then each sample is assumed to be an individual diploid, i.e., neither a polyploid nor a pool.
Optional filtering steps based on minimum depth, minimum allele frequency, and maximum sparsity are available.
Genotype data are not imported into R, LD estimation and imputation per se are multi-threaded, and imputation output is written into disk as an allele frequency table.
The structs, traits, methods, and functions defined in this library are subsets of \href{https://github.com/jeffersonfparil/poolgen}{poolgen}, and will eventually be merged.
Genotype file formats:
\itemize{
\item vcf: canonical variant calling or genotype data format for individual samples. This should include the \code{AD} field (allele depth), and may or may not have genotypes called (e.g. generated via bctools mpileup -a AD,DP ...). If the \code{GT} field is present but the \code{AD} field is absent, then each sample is assumed to be an individual diploid, i.e., neither a polyploid nor a pool. The \href{#vcf2sync}{\code{vcf2sync}} utility is expected to work with vcf versions 4.2 and 4.3. See \href{https://samtools.github.io/hts-specs/VCFv4.2.pdf}{VCFv4.2} and \href{https://samtools.github.io/hts-specs/VCFv4.3.pdf}{VCFv4.3} for details in the format specifications.
\item sync: an extension of \href{https://academic.oup.com/bioinformatics/article/27/24/3435/306737}{popoolation2's} sync or synchronised pileup file format, which includes a header line prepended by '#' showing the names of each column including the names of each pool. Additional header line/s and comments prepended with '#' may be added anywhere within the file.
\itemize{
\item tab-delimited
\item \emph{Header line/s}:  optional header line/s including the names of the pools, e.g. \verb{# chr pos ref pool1 pool2 pool3 pool4 pool5}
\item \emph{Column 1}:       chromosome or scaffold name
\item \emph{Column 2}:       locus position
\item \emph{Column 3}:       reference allele, e.g. A, T, C, G
\item \emph{Column/s 4 to n}:  colon-delimited allele counts: A:T:C:G:DEL:N, where "DEL" refers to insertion/deletion, and "N" is unclassified. A pool or population or polyploid individual is represented by a single column of this colon-delimited allele counts.
}
\item allele frequency table
\itemize{
\item comma-delimited
\item \emph{Header line}: \verb{ #chr,pos,allele,<pool_name_1>,...,<pool_name_n>}
\item each locus is represented by 2 or more rows, i.e. 2 for biallelic loci, and >2 for multi-allelic loci
}
}
}
